Adding validation checks to avoid users using the app with invalid account credentials; updating namespaces; other small refactors
