+++ src/App/ChaosRecipeEnhancer.UI/UserControls/SettingsForms/GeneralForms/RecipesForm.xaml
@@ -7,12 +7,13 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:zemotoCommon="clr-namespace:ChaosRecipeEnhancer.UI.Utilities.ZemotoCommon"
     mc:Ignorable="d"
     Background="#202020"
     d:DataContext="{d:DesignInstance Type={x:Type forms:RecipesFormViewModel}}">
 
     <UserControl.Resources>
-        <utilities:InvertBoolConverter x:Key="InvertedBoolConverter" />
+        <zemotoCommon:InvertBoolConverter x:Key="InvertedBoolConverter" />
     </UserControl.Resources>
 
     <Grid
