+++ src/App/ChaosRecipeEnhancer.UI/Utilities/ZemotoCommon/Converters.cs
@@ -6,7 +6,7 @@ using System.Windows;
 using System.Windows.Data;
 using System.Windows.Media;
 
-namespace ChaosRecipeEnhancer.UI.Utilities;
+namespace ChaosRecipeEnhancer.UI.Utilities.ZemotoCommon;
 
 public sealed class BoolVisibilityConverter : IValueConverter
 {
